Move start_announcements_processor to networking
